Moved class to different package
